css/css-conditional/js/CSS-supports.html

<!doctype html>
<meta charset="utf-8">
<title>CSS.supports</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional/#dom-css-supports">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    test(function () {
        // https://drafts.csswg.org/css-conditional/#dom-css-supports
        // https://drafts.csswg.org/css-conditional/#typedef-supports-condition
        assert_equals(CSS.supports("color: red"), true, "CSS.supports: Single-argument form allows for declarations without enclosing parentheses");
        assert_equals(CSS.supports("(color: red) and (color: blue)"), true, "CSS.supports: Complex conditions allowed");
        assert_equals(CSS.supports("not (foobar)"), true, "CSS.supports: general_enclosed still parses");
        assert_equals(CSS.supports("color: something-pointless var(--foo)"), true, "Variable references always parse");
        assert_equals(CSS.supports("color: something-pointless(var(--foo))"), true, "Variable references in an unknown function always parse");
    }, "CSS.supports, one argument form");
    test(function () {
        // https://drafts.csswg.org/css-conditional/#dom-css-supports
        // https://drafts.csswg.org/css-conditional/#dfn-support
        assert_equals(CSS.supports("color", "red"), true, "CSS.supports: two argument form succeeds for known property");
        assert_equals(CSS.supports("unknownproperty", "blah"), false, "CSS.supports: two argument form fails for unknown property");
        assert_equals(CSS.supports("width", "blah"), false, "CSS.supports: two argument form fails for invalid value");
        assert_equals(CSS.supports("--foo", "blah"), true, "CSS.supports: two argument form succeeds for custom property");
    }, "CSS.supports, two argument form");
    test(function () {
        assert_equals(CSS.supports("selector(div)"), true, "CSS.supports: selector() function accepts a selector");
        assert_equals(CSS.supports("selector(div, div)"), false, "CSS.supports: selector() function doesn't accept a selector list");
        assert_equals(CSS.supports("selector(::-webkit-unknown-pseudo-element)"), false, "CSS.supports: selector() function rejects unknown webkit pseudo-elements.");
        assert_equals(CSS.supports("selector(::before)"), true, "CSS.supports: selector() function accepts known pseudo-elements");
        assert_equals(CSS.supports("selector(div + .c)"), true, "CSS.supports: selector() with simple combinators");
        assert_equals(CSS.supports("selector(div | .c)"), false, "CSS.supports: selector() with unknown combinators");
    }, "CSS.supports, selector function");
</script>
